streamlit_max_agent/requirements.txt
streamlit>=1.28.0
openai>=1.3.0
chromadb>=0.4.15
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
pypdf2>=3.0.1
python-docx>=1.1.0
tiktoken>=0.5.1
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0